requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
## The following requirements were added by pip freeze:
black==23.1.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
python-dotenv==1.0.0
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
tomli==2.0.1
